+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Until May 2022 (inclusive) no changelog was kept. We might try to reconstruct it
 * Housekeeping: Update README with information relevant to SHiP
 * GEANT4: Update version to v11.2.1
 * flatbuffers: Update recipe from ALICE
+* BOOST: update recipe and version to 1.85
 
 ### Removed